requirements.txt
django
psycopg2
gunicorn
GitPython